MeOS V4.0 Update 2

+++ code/onlineresults.h
@@ -61,7 +61,7 @@ protected:
 
   wstring getExportFileName() const;
   int bytesExported;
-  DWORD lastSync;
+  uint64_t lastSync;
 
   vector<string> errorLines;
   void formatError(gdioutput &gdi);
